Conflicts:
	CHANGELOG
	README
	README.rst
	bin/filebucket
	bin/pi
	bin/puppet
	bin/puppetdoc
	bin/ralsh
	conf/redhat/puppet.spec
	conf/suse/puppet.spec
	ext/nagios/check_puppet.rb
	ext/regexp_nodes/regexp_nodes.rb
	install.rb
	lib/puppet.rb
	lib/puppet/application/puppetd.rb
	lib/puppet/configurer.rb
	lib/puppet/defaults.rb
	lib/puppet/file_serving/fileset.rb
	lib/puppet/file_serving/terminus_helper.rb
	lib/puppet/indirector/indirection.rb
	lib/puppet/network/authstore.rb
	lib/puppet/network/format_handler.rb
	lib/puppet/network/handler/fileserver.rb
	lib/puppet/network/handler/master.rb
	lib/puppet/network/http_server/webrick.rb
	lib/puppet/parser/ast/casestatement.rb
	lib/puppet/parser/ast/ifstatement.rb
	lib/puppet/parser/ast/leaf.rb
	lib/puppet/parser/ast/selector.rb
	lib/puppet/parser/functions/generate.rb
	lib/puppet/parser/functions/require.rb
	lib/puppet/parser/functions/template.rb
	lib/puppet/parser/lexer.rb
	lib/puppet/parser/resource.rb
	lib/puppet/parser/scope.rb
	lib/puppet/provider/augeas/augeas.rb
	lib/puppet/provider/macauthorization/macauthorization.rb
	lib/puppet/provider/maillist/mailman.rb
	lib/puppet/provider/package/portage.rb
	lib/puppet/provider/service/debian.rb
	lib/puppet/provider/service/smf.rb
	lib/puppet/provider/ssh_authorized_key/parsed.rb
	lib/puppet/reference/configuration.rb
	lib/puppet/reference/providers.rb
	lib/puppet/resource/catalog.rb
	lib/puppet/simple_graph.rb
	lib/puppet/ssl/certificate.rb
	lib/puppet/ssl/certificate_authority.rb
	lib/puppet/ssl/host.rb
	lib/puppet/sslcertificates/ca.rb
	lib/puppet/transaction.rb
	lib/puppet/type/augeas.rb
	lib/puppet/type/exec.rb
	lib/puppet/type/file.rb
	lib/puppet/type/file/checksum.rb
	lib/puppet/type/file/source.rb
	lib/puppet/type/maillist.rb
	lib/puppet/type/mount.rb
	lib/puppet/type/tidy.rb
	lib/puppet/util/checksums.rb
	lib/puppet/util/monkey_patches.rb
	lib/puppet/util/nagios_maker.rb
	lib/puppet/util/settings.rb
	lib/puppet/util/suidmanager.rb
	lib/puppet/util/zaml.rb
	man/man8/puppetca.8
	man/man8/puppetd.8
	man/man8/puppetmasterd.8
	man/man8/puppetqd.8
	man/man8/puppetrun.8
	sbin/puppetca
	sbin/puppetd
	sbin/puppetmasterd
	sbin/puppetqd
	sbin/puppetrun
	spec/integration/defaults.rb
	spec/integration/parser/functions/require.rb
	spec/unit/application/filebucket.rb
	spec/unit/application/puppetd.rb
	spec/unit/configurer.rb
	spec/unit/file_serving/fileset.rb
	spec/unit/file_serving/terminus_helper.rb
	spec/unit/indirector/indirection.rb
	spec/unit/network/format_handler.rb
	spec/unit/parser/ast/casestatement.rb
	spec/unit/parser/ast/ifstatement.rb
	spec/unit/parser/ast/leaf.rb
	spec/unit/parser/ast/selector.rb
	spec/unit/parser/functions/require.rb
	spec/unit/parser/resource.rb
	spec/unit/parser/scope.rb
	spec/unit/provider/augeas/augeas.rb
	spec/unit/provider/ssh_authorized_key/parsed.rb
	spec/unit/resource/catalog.rb
	spec/unit/simple_graph.rb
	spec/unit/ssl/host.rb
	spec/unit/type/file.rb
	spec/unit/type/tidy.rb
	spec/unit/util/checksums.rb
	spec/unit/util/settings.rb
	test/certmgr/certmgr.rb
	test/network/handler/ca.rb
	test/network/handler/fileserver.rb
	test/other/transactions.rb
	test/puppet/tc_suidmanager.rb
	test/ral/providers/package.rb
	test/ral/type/file.rb
	test/ral/type/resources.rb
	test/util/settings.rb

+++ lib/puppet/rails/database/003_add_environment_to_host.rb
@@ -1,13 +1,9 @@
 class AddEnvironmentToHost < ActiveRecord::Migration
-    def self.up
-        unless ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
-            add_column :hosts, :environment, :string
-        end
-    end
+  def self.up
+    add_column :hosts, :environment, :string unless ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
+  end
 
-    def self.down
-        if ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
-            remove_column :hosts, :environment
-        end
-    end
+  def self.down
+    remove_column :hosts, :environment if ActiveRecord::Base.connection.columns(:hosts).collect {|c| c.name}.include?("environment")
+  end
 end
